src/requirements.txt
# src/requirements.txt
# Core dependencies
numpy>=1.19.0
pandas>=1.1.0
pydantic>=1.8.0
matplotlib>=3.3.0
scikit-learn>=0.24.0
scipy>=1.6.0
umap-learn>=0.5.1
seaborn>=0.11.0
opencv-python>=4.5.0
scikit-image>=0.18.0 